requirements.txt
torch >= 2.0.1
torchtext==0.15.2
torchdata==0.6.1
matplotlib >= 3.7.1   
tqdm >= 4.66.1        
numpy >= 1.25, < 2.0  
pandas >= 2.2.1       
tiktoken==0.7.0
pyyaml==6.0.1